LPC4330 Xplorer: Rename files for better conformance to naming conventions
