+++ nuttx/configs/lpc4330-xplorer/src/Makefile
@@ -35,64 +35,64 @@
 
 -include $(TOPDIR)/Make.defs
 
-CFLAGS		+= -I$(TOPDIR)/sched
+CFLAGS += -I$(TOPDIR)/sched
 
-ASRCS		=
-CSRCS		= up_boot.c
+ASRCS =
+CSRCS = lpc43_boot.c
 
 ifeq ($(CONFIG_NSH_ARCHINIT),y)
-CSRCS		+= up_nsh.c
+CSRCS += lpc43_nsh.c
 endif
 
 ifeq ($(CONFIG_ARCH_FPU),y)
-CSRCS		+= up_ostest.c
+CSRCS += lpc43_ostest.c
 endif
 
 ifeq ($(CONFIG_ARCH_LEDS),y)
-CSRCS		+= up_autoleds.c
+CSRCS += lpc43_autoleds.c
 else
-CSRCS		+= up_userleds.c
+CSRCS += lpc43_userleds.c
 endif
 
 ifeq ($(CONFIG_ARCH_BUTTONS),y)
-CSRCS		+= up_buttons.c
+CSRCS += lpc43_buttons.c
 endif
 
 ifeq ($(CONFIG_USBMSC),y)
-CSRCS		+= up_usbmsc.c
+CSRCS += up_usbmsc.c
 endif
 
 ifeq ($(CONFIG_SPIFI_LIBRARY),y)
-CFLAGS		+= -DEUROBIRD
-SPIFI_LIB	= spifi_lib
-CSRCS		+= $(SPIFI_LIB)/spifi_rom_api.c
-#CSRCS		+= $(SPIFI_LIB)/amic.c
-#CSRCS		+= $(SPIFI_LIB)/atmel.c
-#CSRCS		+= $(SPIFI_LIB)/chi.c
-#CSRCS		+= $(SPIFI_LIB)/eon.c
-#CSRCS		+= $(SPIFI_LIB)/esmt.c
-#CSRCS		+= $(SPIFI_LIB)/esmt.c
-#CSRCS		+= $(SPIFI_LIB)/giga.c
-#CSRCS		+= $(SPIFI_LIB)/macronix.c
-#CSRCS		+= $(SPIFI_LIB)/numonyx.c
-CSRCS		+= $(SPIFI_LIB)/spansion.c
-#CSRCS		+= $(SPIFI_LIB)/sst.c
-CSRCS		+= $(SPIFI_LIB)/winbond.c
+CFLAGS += -DEUROBIRD
+SPIFI_LIB = spifi_lib
+CSRCS += $(SPIFI_LIB)/spifi_rom_api.c
+#CSRCS += $(SPIFI_LIB)/amic.c
+#CSRCS += $(SPIFI_LIB)/atmel.c
+#CSRCS += $(SPIFI_LIB)/chi.c
+#CSRCS += $(SPIFI_LIB)/eon.c
+#CSRCS += $(SPIFI_LIB)/esmt.c
+#CSRCS += $(SPIFI_LIB)/esmt.c
+#CSRCS += $(SPIFI_LIB)/giga.c
+#CSRCS += $(SPIFI_LIB)/macronix.c
+#CSRCS += $(SPIFI_LIB)/numonyx.c
+CSRCS += $(SPIFI_LIB)/spansion.c
+#CSRCS += $(SPIFI_LIB)/sst.c
+CSRCS += $(SPIFI_LIB)/winbond.c
 endif
 
-AOBJS		= $(ASRCS:.S=$(OBJEXT))
-COBJS		= $(CSRCS:.c=$(OBJEXT))
+AOBJS = $(ASRCS:.S=$(OBJEXT))
+COBJS = $(CSRCS:.c=$(OBJEXT))
 
-SRCS		= $(ASRCS) $(CSRCS)
-OBJS		= $(AOBJS) $(COBJS)
+SRCS = $(ASRCS) $(CSRCS)
+OBJS = $(AOBJS) $(COBJS)
 
-ARCH_SRCDIR	= $(TOPDIR)/arch/$(CONFIG_ARCH)/src
+ARCH_SRCDIR = $(TOPDIR)/arch/$(CONFIG_ARCH)/src
 ifeq ($(WINTOOL),y)
-  CFLAGS	+= -I "${shell cygpath -w $(ARCH_SRCDIR)/chip}" \
-		   -I "${shell cygpath -w $(ARCH_SRCDIR)/common}" \
-		   -I "${shell cygpath -w $(ARCH_SRCDIR)/armv7-m}"
+  CFLAGS += -I "${shell cygpath -w $(ARCH_SRCDIR)/chip}" \
+            -I "${shell cygpath -w $(ARCH_SRCDIR)/common}" \
+            -I "${shell cygpath -w $(ARCH_SRCDIR)/armv7-m}"
 else
-  CFLAGS	+= -I$(ARCH_SRCDIR)/chip -I$(ARCH_SRCDIR)/common -I$(ARCH_SRCDIR)/armv7-m
+  CFLAGS += -I$(ARCH_SRCDIR)/chip -I$(ARCH_SRCDIR)/common -I$(ARCH_SRCDIR)/armv7-m
 endif
 
 all: libboard$(LIBEXT)
