+++ nuttx/configs/lpc4330-xplorer/src/lpc43_boot.c
@@ -1,5 +1,5 @@
 /************************************************************************************
- * configs/lpc4330-xplorer/src/up_boot.c
+ * configs/lpc4330-xplorer/src/lpc43_boot.c
  *
  *   Copyright (C) 2012, 2015 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -50,7 +50,7 @@
 #include "xplorer_internal.h"
 
 /************************************************************************************
- * Definitions
+ * Pre-processor Definitions
  ************************************************************************************/
 
 /************************************************************************************
